requirements.txt
docxcompose==1.3.3
docxtpl==0.12.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.21.2
pandas==1.3.3
Pillow==8.3.2
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2021.1
six==1.16.0
